acme: fix X, Y command crashing with extra locks
When mouse issues a command, the window that has the mouse is locked
with winlock, then the command is run, and then winunlock.  Runpipe
is coping this mechanism by sequentially calling incref, run (which
contains a winclose that does a decref, winunlock, and finally
winlock again.  This is fine for the usual pipe_cmd, but for X_cmd,
all except one window does not have the lock to begin with.  Acme
would crash when such X, Y command were run and winunlock were
called on a window without a prior winlock.

This patch fix the issue by adding extra winunlock/winlock in X_cmd
to deal with existing locks, and applying winlock/winunlock for each
matching window in filelooper.